some changes to find issue with JNA 4

+++ contrib/platform/src/com/sun/jna/platform/win32/COM/tlb/imp/TlbFunction.java
@@ -52,8 +52,8 @@ public class TlbFunction extends TlbAbstractMethod implements Variant {
 
         String methodparams = "";
         String methodvariables = "";
-        short vtableId = funcDesc.oVft;
-        short paramCount = funcDesc.cParams;
+        int vtableId = funcDesc.oVft;
+        int paramCount = funcDesc.cParams;
         ELEMDESC elemDesdRetType = funcDesc.elemdescFunc;
         String returnType = this.getType(elemDesdRetType.tdesc);
         String[] names = typeInfoUtil.getNames(funcDesc.memid, paramCount + 1);
